+++ src/impls_for_tuple.rs
@@ -1,12 +1,12 @@
-use hecs::Query;
+use hecs::{Component, Entity, Query};
 use resources::{Resource, Resources};
 
-#[cfg(feature = "parallel")]
 use hecs::World;
 
 use crate::{
+    fetch_components::{ComponentEffector, Fetch as ComponentFetch, Mutability, Optionality},
     query_bundle::{QueryBundle, QueryEffector, QuerySingle, QueryUnit},
-    resource_bundle::{Fetch, Mutability, ResourceBundle, ResourceEffector, ResourceSingle},
+    resource_bundle::{Fetch as ResourceFetch, ResourceBundle, ResourceEffector, ResourceSingle},
     system::TupleAppend,
 };
 
@@ -38,10 +38,15 @@ where
     Q: QueryUnit,
     Self: Query,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = Q::ComponentEffector;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Q::component_effector()
     }
 
     #[cfg(feature = "parallel")]
@@ -59,10 +64,10 @@ impl<Q> QueryBundle for (Q,)
 where
     Q: QuerySingle,
 {
-    type Effectors = Q::Effector;
+    type QueryEffectors = Q::QueryEffector;
 
-    fn effectors() -> Self::Effectors {
-        Q::effector()
+    fn query_effectors() -> Self::QueryEffectors {
+        Q::query_effector()
     }
 
     #[cfg(feature = "parallel")]
@@ -100,19 +105,41 @@ macro_rules! impls_for_tuple {
         }
 
         paste::item! {
-            impl<'a, $([<M $letter>]),*, $([<R $letter>]),*> Fetch<'a>
+            impl<'a, $([<M $letter>]),*, $([<R $letter>]),*> ResourceFetch<'a>
                 for ($(ResourceEffector<[<M $letter>], [<R $letter>]>,)*)
             where
                 $([<M $letter>]: Mutability,)*
                 $([<R $letter>]: Resource,)*
-                $(ResourceEffector<[<M $letter>], [<R $letter>]>: Fetch<'a>,)*
+                $(ResourceEffector<[<M $letter>], [<R $letter>]>: ResourceFetch<'a>,)*
             {
                 type Refs = (
-                    $(<ResourceEffector<[<M $letter>], [<R $letter>]> as Fetch<'a>>::Refs,)*
+                    $(<ResourceEffector<[<M $letter>], [<R $letter>]> as ResourceFetch<'a>>::Refs,)*
                 );
 
                 fn fetch(&self, resources: &'a Resources) -> Self::Refs {
                     ($(ResourceEffector::<[<M $letter>], [<R $letter>]>::new().fetch(resources),)*)
+                }
+            }
+        }
+
+        paste::item! {
+            impl<'a, $([<M $letter>]),*, $([<O $letter>]),*, $([<C $letter>]),*> ComponentFetch<'a>
+                for ($(ComponentEffector<[<M $letter>], [<O $letter>], [<C $letter>]>,)*)
+            where
+                $([<M $letter>]: Mutability,)*
+                $([<O $letter>]: Optionality,)*
+                $([<C $letter>]: Component,)*
+                $(ComponentEffector<[<M $letter>], [<O $letter>], [<C $letter>]>:
+                    ComponentFetch<'a>,)*
+            {
+                type Refs = (
+                    $(<ComponentEffector<[<M $letter>], [<O $letter>], [<C $letter>]>
+                        as ComponentFetch<'a>>::Refs,)*
+                );
+
+                fn fetch(&self, world: &'a World, entity: Entity) -> Self::Refs {
+                    ($(ComponentEffector::<[<M $letter>], [<O $letter>], [<C $letter>]>::new()
+                        .fetch(world, entity),)*)
                 }
             }
         }
@@ -122,10 +149,15 @@ macro_rules! impls_for_tuple {
             $($letter: QueryUnit,)*
             Self: Query,
         {
-            type Effector = QueryEffector<Self>;
+            type QueryEffector = QueryEffector<Self>;
+            type ComponentEffectors = ($($letter::ComponentEffector,)*);
 
-            fn effector() -> Self::Effector {
+            fn query_effector() -> Self::QueryEffector {
                 QueryEffector::new()
+            }
+
+            fn component_effectors() -> Self::ComponentEffectors {
+                ($($letter::component_effector(),)*)
             }
 
             #[cfg(feature = "parallel")]
@@ -143,10 +175,10 @@ macro_rules! impls_for_tuple {
         where
             $($letter: Query + QuerySingle + Send + Sync,)*
         {
-            type Effectors = ($($letter::Effector,)*);
+            type QueryEffectors = ($($letter::QueryEffector,)*);
 
-            fn effectors() -> Self::Effectors {
-                ($($letter::effector(),)*)
+            fn query_effectors() -> Self::QueryEffectors {
+                ($($letter::query_effector(),)*)
             }
 
             #[cfg(feature = "parallel")]