+++ src/query_bundle.rs
@@ -1,10 +1,12 @@
-use hecs::{Component, Query, QueryBorrow, With, Without, World};
+use hecs::{Component, Query, With, Without};
 use std::marker::PhantomData;
 
 #[cfg(feature = "parallel")]
-use hecs::Access;
+use hecs::{Access, World};
 #[cfg(feature = "parallel")]
 use std::any::TypeId;
+
+use crate::fetch_components::{ComponentEffector, Immutable, Mandatory, Mutable, Optional};
 
 #[cfg(feature = "parallel")]
 use crate::{ArchetypeAccess, SystemBorrows};
@@ -20,23 +22,10 @@ impl<Q> QueryEffector<Q>
 where
     Q: Query + Send + Sync,
 {
-    pub fn new() -> Self {
+    pub(crate) fn new() -> Self {
         Self {
             phantom_data: PhantomData,
         }
-    }
-
-    pub fn query<'a>(&self, world: &'a World) -> QueryBorrow<'a, Q> {
-        world.query::<Q>()
-    }
-}
-
-impl<Q> Default for QueryEffector<Q>
-where
-    Q: Query + Send + Sync,
-{
-    fn default() -> Self {
-        QueryEffector::new()
     }
 }
 
@@ -52,14 +41,21 @@ where
 impl<Q> Copy for QueryEffector<Q> where Q: Query + Send + Sync {}
 
 pub trait QueryUnit: Send + Sync {
+    type ComponentEffector;
+
+    fn component_effector() -> Self::ComponentEffector;
+
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows);
 }
 
 pub trait QuerySingle: Send + Sync {
-    type Effector;
+    type QueryEffector;
+    type ComponentEffectors;
 
-    fn effector() -> Self::Effector;
+    fn query_effector() -> Self::QueryEffector;
+
+    fn component_effectors() -> Self::ComponentEffectors;
 
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows);
@@ -69,9 +65,9 @@ pub trait QuerySingle: Send + Sync {
 }
 
 pub trait QueryBundle: Send + Sync {
-    type Effectors;
+    type QueryEffectors;
 
-    fn effectors() -> Self::Effectors;
+    fn query_effectors() -> Self::QueryEffectors;
 
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows);
@@ -84,6 +80,12 @@ impl<C> QueryUnit for &'_ C
 where
     C: Component,
 {
+    type ComponentEffector = ComponentEffector<Immutable, Mandatory, C>;
+
+    fn component_effector() -> Self::ComponentEffector {
+        ComponentEffector::new()
+    }
+
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows) {
         borrows.components_immutable.insert(TypeId::of::<C>());
@@ -94,26 +96,57 @@ impl<C> QueryUnit for &'_ mut C
 where
     C: Component,
 {
+    type ComponentEffector = ComponentEffector<Mutable, Mandatory, C>;
+
+    fn component_effector() -> Self::ComponentEffector {
+        ComponentEffector::new()
+    }
+
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows) {
         borrows.components_mutable.insert(TypeId::of::<C>());
     }
 }
 
-impl<Q> QueryUnit for Option<Q>
+impl<C> QueryUnit for Option<&'_ C>
 where
-    Q: QueryUnit,
+    C: Component,
 {
+    type ComponentEffector = ComponentEffector<Immutable, Optional, C>;
+
+    fn component_effector() -> Self::ComponentEffector {
+        ComponentEffector::new()
+    }
+
     #[cfg(feature = "parallel")]
     fn write_borrows(borrows: &mut SystemBorrows) {
-        Q::write_borrows(borrows);
+        borrows.components_immutable.insert(TypeId::of::<C>());
+    }
+}
+
+impl<C> QueryUnit for Option<&'_ mut C>
+where
+    C: Component,
+{
+    type ComponentEffector = ComponentEffector<Mutable, Optional, C>;
+
+    fn component_effector() -> Self::ComponentEffector {
+        ComponentEffector::new()
+    }
+
+    #[cfg(feature = "parallel")]
+    fn write_borrows(borrows: &mut SystemBorrows) {
+        borrows.components_mutable.insert(TypeId::of::<C>());
     }
 }
 
 impl QuerySingle for () {
-    type Effector = ();
+    type QueryEffector = ();
+    type ComponentEffectors = ();
 
-    fn effector() -> Self::Effector {}
+    fn query_effector() -> Self::QueryEffector {}
+
+    fn component_effectors() -> Self::ComponentEffectors {}
 
     #[cfg(feature = "parallel")]
     fn write_borrows(_: &mut SystemBorrows) {}
@@ -127,10 +160,15 @@ where
     C: Component,
     Self: Query + QueryUnit,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = <Self as QueryUnit>::ComponentEffector;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Self::component_effector()
     }
 
     #[cfg(feature = "parallel")]
@@ -149,10 +187,15 @@ where
     C: Component,
     Self: Query + QueryUnit,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = <Self as QueryUnit>::ComponentEffector;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Self::component_effector()
     }
 
     #[cfg(feature = "parallel")]
@@ -171,10 +214,15 @@ where
     Q: QueryUnit,
     Self: Query + QueryUnit,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = <Self as QueryUnit>::ComponentEffector;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Self::component_effector()
     }
 
     #[cfg(feature = "parallel")]
@@ -193,10 +241,15 @@ where
     C: Component,
     Q: Query + QuerySingle,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = Q::ComponentEffectors;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Q::component_effectors()
     }
 
     #[cfg(feature = "parallel")]
@@ -215,10 +268,15 @@ where
     C: Component,
     Q: Query + QuerySingle,
 {
-    type Effector = QueryEffector<Self>;
+    type QueryEffector = QueryEffector<Self>;
+    type ComponentEffectors = Q::ComponentEffectors;
 
-    fn effector() -> Self::Effector {
+    fn query_effector() -> Self::QueryEffector {
         QueryEffector::new()
+    }
+
+    fn component_effectors() -> Self::ComponentEffectors {
+        Q::component_effectors()
     }
 
     #[cfg(feature = "parallel")]
@@ -233,9 +291,9 @@ where
 }
 
 impl QueryBundle for () {
-    type Effectors = ();
+    type QueryEffectors = ();
 
-    fn effectors() -> Self::Effectors {}
+    fn query_effectors() -> Self::QueryEffectors {}
 
     #[cfg(feature = "parallel")]
     fn write_borrows(_: &mut SystemBorrows) {}